[SERVER][MODIF + AJOUT] ajout class decor + inplementation parser de level

+++ includes/server/Map.h
@@ -11,7 +11,7 @@ struct			t_spawn
 {
   entityType		type;
   unsigned short int	y;
-  unsigned int		time;
+  unsigned long int		time;
 };
 
 class			Map
